Refactor SystemCrawler to allow passing filters (so we can pass new filters for a "live search" in the editor)
