+++ app/src/main/java/eu/darken/sdmse/systemcleaner/core/SystemCleaner.kt
@@ -24,6 +24,7 @@ import eu.darken.sdmse.exclusion.core.types.PathExclusion
 import eu.darken.sdmse.exclusion.core.types.excludeNestedLookups
 import eu.darken.sdmse.main.core.SDMTool
 import eu.darken.sdmse.systemcleaner.core.filter.FilterIdentifier
+import eu.darken.sdmse.systemcleaner.core.filter.FilterSource
 import eu.darken.sdmse.systemcleaner.core.tasks.SystemCleanerDeleteTask
 import eu.darken.sdmse.systemcleaner.core.tasks.SystemCleanerScanTask
 import eu.darken.sdmse.systemcleaner.core.tasks.SystemCleanerSchedulerTask
@@ -33,7 +34,6 @@ import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import javax.inject.Inject
-import javax.inject.Provider
 import javax.inject.Singleton
 
 @Singleton
@@ -41,8 +41,9 @@ class SystemCleaner @Inject constructor(
     @AppScope private val appScope: CoroutineScope,
     fileForensics: FileForensics,
     private val gatewaySwitch: GatewaySwitch,
-    private val crawlerProvider: Provider<SystemCrawler>,
+    private val crawler: SystemCrawler,
     private val exclusionManager: ExclusionManager,
+    private val filterSource: FilterSource,
     pkgOps: PkgOps,
 ) : SDMTool, Progress.Client {
 
@@ -93,10 +94,8 @@ class SystemCleaner @Inject constructor(
 
         internalData.value = null
 
-        val crawler = crawlerProvider.get()
-
         val results = crawler.withProgress(this) {
-            crawl()
+            crawl(filterSource.create())
         }
 
         log(TAG) { "Warming up fields..." }
